Refactored LoginControllerTest and LoginTest
Now UserControllerTest and UserTest
Updated UserTest; added tests

Added comments to both tests
